+++ flake8_raise.py
@@ -1,10 +1,6 @@
 import ast
 
 
-class Context(list):
-    allow_bare_raise = True
-
-
 class RaiseStatementChecker:
     name = "flake8-raise"
     version = "0.0.4"
@@ -15,16 +11,15 @@ class RaiseStatementChecker:
 
     def __init__(self, tree, filename):
         self.tree = tree
-        self.handlers = [Context()]
+        self.handlers = [[]]
 
     def run(self):
         yield from self.walk(self.tree)
 
     def walk(self, node):
         if isinstance(node, ast.FunctionDef):
-            self.handlers.append(Context())
+            self.handlers.append([])
         elif isinstance(node, ast.ExceptHandler):
-            self.handlers[-1].allow_bare_raise = True
             self.handlers[-1].append(node.name)
         elif isinstance(node, ast.Raise):
             if self.handlers[-1] and node.exc is not None:
@@ -34,8 +29,7 @@ class RaiseStatementChecker:
                     yield node.lineno, node.col_offset, self.text["R100"], type(self)
 
                 if (
-                    self.handlers[-1].allow_bare_raise
-                    and isinstance(node.exc, ast.Name)
+                    isinstance(node.exc, ast.Name)
                     and node.exc.id == self.handlers[-1][-1]
                 ):
                     yield node.lineno, node.col_offset, self.text["R101"], type(self)
@@ -46,5 +40,4 @@ class RaiseStatementChecker:
         if isinstance(node, ast.FunctionDef):
             del self.handlers[-1]
         elif isinstance(node, ast.ExceptHandler):
-            self.handlers[-1].allow_bare_raise = False
             del self.handlers[-1][-1]