+++ tests/data/reraise_exception_after_other_except_bare_raise.py
@@ -1,0 +1,8 @@
+try:
+    pass
+except ValueError:
+    try:
+        pass
+    except TypeError:
+        pass
+    raise