+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ UNRELEASED
 
 - Include ``CHANGELOG.rst``, ``LICENSE``, ``tox.ini``, and ``tests`` in the
   source distribution.
+- Improved "R101 Use bare raise in except handler" for Python 3 nested
+  ``except`` handlers.
 
 0.0.4 (2020-01-19)
 ------------------